+++ CHT8310.h
@@ -14,10 +14,12 @@
 
 #define CHT8310_LIB_VERSION              (F("0.1.1"))
 
+
 //  DEFAULT ADDRESS
 #ifndef CHT8310_DEFAULT_ADDRESS
 #define CHT8310_DEFAULT_ADDRESS          0x40
 #endif
+
 
 //  ERRORS
 #define CHT8310_OK                       0
@@ -26,26 +28,6 @@
 #define CHT8310_ERROR_CONNECT            -12
 #define CHT8310_ERROR_LASTREAD           -20
 #define CHT8310_ERROR_HUMIDITY           -30
-
-
-//  REGISTERS
-#define CHT8310_REG_TEMPERATURE          0x00
-#define CHT8310_REG_HUMIDITY             0x01
-
-#define CHT8310_REG_SWRESET              0xFC
-#define CHT8310_REG_MANUFACTURER         0xFF
-
-
-//  REGISTER MASKS
-//  not implemented in 0.1.0
-#define CHT8310_REG_STATUS               0x02
-#define CHT8310_REG_CONFIG               0x03
-#define CHT8310_REG_CONVERT_RATE         0x04
-#define CHT8310_REG_TEMP_HIGH_LIMIT      0x05
-#define CHT8310_REG_TEMP_LOW_LIMIT       0x06
-#define CHT8310_REG_HUM_HIGH_LIMIT       0x07
-#define CHT8310_REG_HUM_LOW_LIMIT        0x08
-#define CHT8310_REG_ONESHOT              0x0F
 
 
 class CHT8310
@@ -84,18 +66,29 @@ public:
   float    getTemperatureOffset();
 
 
+  //  CONVERT RATE (not tested, under development)
+  void     setConvertRate(uint8_t rate = 4);
+  uint8_t  getConvertRate();
+
+
   //  ALERT (not tested, under development)
   void     setTemperatureHighLimit(float temperature);
   void     setTemperatureLowLimit(float temperature);
   void     setHumidityHighLimit(float humidity);
   void     setHumidityLowLimit(float humidity);
 
+
   //  STATUS (not tested, under development)
   uint16_t getStatusRegister();
 
 
+  //  SOFTWARE RESET (not tested, under development)
+  void     softwareReset();
+
+
   //  META DATA
   uint16_t getManufacturer();     //  expect 0x5959
+
 
   //  PATCH TO ACCESS REGISTERS
   uint16_t readRegister(uint8_t reg);